% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycparks.R
\name{nycparks_art}
\alias{nycparks_art}
\title{Fetch NYC Parks temporary public art installations}
\usage{nycparks_art()}
\value{tibble with art installation details}
\description{Directory of temporary art exhibitions since 2000.}
