% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycparks.R
\name{nycparks_closures}
\alias{nycparks_closures}
\title{Fetch current NYC Parks closure notifications}
\usage{nycparks_closures()}
\value{tibble with park/facility closure details}
\description{Fetch current NYC Parks closure notifications.}
