% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycparks.R
\name{nycparks_facilities}
\alias{nycparks_facilities}
\title{Fetch a NYC Parks facility directory}
\usage{nycparks_facilities(directory, borough = NULL)}
\arguments{
\item{directory}{Directory name (see nycparks_list())}
\item{borough}{Optional borough filter}
}
\value{tibble with facility details}
\description{Retrieves a full directory of a specific facility type.}
