% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycparks.R
\name{nycparks_history}
\alias{nycparks_history}
\title{Fetch NYC Parks historical signs}
\usage{nycparks_history(borough = NULL, query = NULL)}
\arguments{
\item{borough}{Optional borough filter}
\item{query}{Optional text search}
}
\value{tibble with historical sign data}
\description{Returns the full text of historical signs in NYC Parks.}
