% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nycparks.R
\name{nycparks_list}
\alias{nycparks_list}
\title{List available NYC Parks facility directories}
\usage{nycparks_list()}
\value{character vector of directory names}
\description{Returns the names of all known directory feeds.}
